--- public/js/api-params.js
@@ -1,6 +1,0 @@
-require('dotenv').config();
-
-const baseUrl = 'https://finnhub.io/api/v1/';
-const apiKey = `&token=${process.env.API_KEY}`;
-
-module.exports = { baseUrl, apiKey }
