--- public/js/axios-all.js
@@ -1,24 +1,0 @@
-const axios = require("axios");
-const apiKey 
-
-
-async function getAllNyse() {
-
-
-
-    const response = await axios.get()
-
-}
-getAllNyse();
-
-
-
-
-async function getAllNxas() {
-
-
-
-    const response = await axios.get()
-
-}
-getAllNxas();
